Add instagram endpoint.
Move primary endpoint from taxi to geoapp.  Since this won't just be taxi data, move the endpoint for getting taxi data from GET /taxi to GET /geoapp/taxi.

Refactored to make more code reusable for non-taxi classes.

Add sensible search string parsing.  Support grouping with ( ), negation with - or !, logical or with |, and exact matching or avoidance of string quoted with " and of hashtags.
