+++ conf/geoapp.dist.cfg
@@ -25,3 +25,6 @@ mongo12r: {"order": 4, "name": "Mongo 1/12 Shuffled", "class": "TaxiViaMongoRand
 mongo12: {"order": 5, "name": "Mongo 1/12 Random Value", "class": "TaxiViaMongoCompact", "params": {"dbUri": "mongodb://parakon:27017/taxi12"}}
 mongo: {"order": 6, "name": "Mongo 1/12 Dates", "class": "TaxiViaMongo", "params": {"dbUri": "mongodb://parakon:27017/taxi"}}
 tangelo: {"name": "Tangelo Service 1/12", "class": "TaxiViaTangelo"}
+
+[instagramdata]
+postgres: {"order": 0, "name": "Postgres", "class": "InstagramViaPostgres", "params": {"db": "instagram", "host": "parakon", "user": "taxi", "password": "taxi#1"}}
